+++ R/build-manual.R
@@ -11,11 +11,16 @@ build_manual <- function(pkg = ".", path = NULL) {
   pkg <- as.package(pkg)
   path <- path %||% dirname(pkg$path)
   name <- paste0(pkg$package, "_", pkg$version, ".pdf", collapse = " ")
-  msg <- callr::rcmd("Rd2pdf", cmdargs = c(
+  tryCatch(msg <- callr::rcmd("Rd2pdf", cmdargs = c(
     "--force",
     paste0("--output=", path, "/", name),
     pkg$path
-  ))
+  ), fail_on_status = TRUE),
+  error = function(e) {
+    cat(e$stderr)
+    stop("Failed to build manual", call. = FALSE)
+  })
+
   cat(msg$stdout)
   invisible(msg)
 }