+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `check_rhub` gains a new argument `build_args` for arguments passed to 
 `R CMD build`. `...` is now passed to `rhub::check_for_cran()` (@gaborcsardi, @maelle, #2041)
+* `build_manual()` now fails if the manual fails to build. (#2056)
 
 * `test_file()` and `test_coverage_file()` now work with C
   and C++ files in the src/ directory as well.